Move MultiStrainTB into resistance/multistrain_tb.py.
Keep base TB in tb.py without resistance imports; strain overlay is opt-in via the subpackage.
